requirements.txt
django==4.2
djangorestframework==3.14
reportlab==4.0.8
openpyxl==3.1.2
celery==5.3.6
redis==5.0.6